requirements.txt
Django>=2.1.5
django-cleanup>=3.0.1
djangorestframework>=3.9.1
Markdown>=3.0.1
pytz>=2018.9
six>=1.12.0
psycopg2-binary>=2.7.7
django-rest-auth>=0.9.3
django-oauth-toolkit>=1.2.0
rest_condition>=1.0.3
Pillow>=4.0.0
